edit mutation delete and update post (#1139)

+++ apps/1CD/instagram/backend/src/resolvers/mutations/post/delete-post.ts
@@ -1,12 +1,16 @@
 import { MutationResolvers } from '../../../generated';
 import { PostModel } from '../../../models/post.model';
 
-export const deletePost: MutationResolvers['deletePost'] = async (_, { _id }) => {
-  const deletePost = await PostModel.findByIdAndDelete(_id);
+export const deletePost: MutationResolvers['deletePost'] = async (_, { _id }, { userId }) => {
+  const findPost = await PostModel.findById({ _id: _id });
 
-  if (!deletePost) {
-    throw new Error('Can not delete post');
+  if (!findPost) throw new Error('Not found post');
+
+  const userID = findPost.user.toString();
+
+  if (userID === userId) {
+    const deletePost = await PostModel.findByIdAndDelete(_id);
+    return deletePost;
   }
-
-  return deletePost;
+  throw new Error('Can not delete post');
 };
